refactor, fuzz: Split descriptor parse harnesses into separate files
